Renaming the "Beta" label to "Editor"
The WikiEditor is the default since a long time now. It does not
make much sense to still call it "Beta". This depends on my other
change to the core preferences because of the new word that needs
to be translated.

Depends on I429c8125b8db789d2756d61249f6cf1fa7f2f51c.

Change-Id: Id8b9a47a9d83dc1acdec46cb6e993deac33cc066