Cleanup borders of wikiEditor
We had some crazy stuff here with negative margins, border boxes and
what not. There were some problems with IE versions, zooming in and out
in pages and with the side by side preview.

This should take care of those. Mostly it moves the borders inside the
elements that are 100%. This was also required because of the
side-by-side preview which was using a dirty trick to hide the border at
the top

Bug: 26828
Change-Id: I3c3466fac53df3e79de6f12f404005e55409fa7c