Prevent style jumping in WikiEditor
Add a new styles module. This module makes sure that the textarea is
already properly styled (jumping lineheight bug) before we start the
JS. Also hide the toolbar before it's styles have finished loading, so
you don't see it being constructed on the fly. Move inline styling for
the old toolbar into this stylesheet as well.

Bug: 47708
Change-Id: I40bec0a24dbd295db301e7faefb93d8991981acb