Revert making installable via composer (and add composer test entry point)
This reverts b8abb9b5f2c6 (a self-merged "experiment") and
1bb0ea4f7654f (Convert entry point to use explicit globals).

The maintainers of this extension do not plan on supporting installation
via composer. This functionality was never included in a stable release
branch, and should not break anyone sticking to stable
branches.

Once merged, I can remove the package from packagist.org

The composer.json is now used to add a "composer test" entry point to run
PHPCS, which this extension badly needs.

Change-Id: I2f3db6addd883d94366a696d5185bf05b2695b11