Follow-up cc5ca3a2f6: now that we sample client-side, send all editingStatsIds to the client
If the server-generated editingStatsId was sampled out, we never sent
it to the client. This made this code confusing and difficult to test,
but it also caused a bug: the client-generated editingStatsId
(for when the user presses back) has no place to go in that case,
and gets dropped on the floor.

Bug: T95919
Change-Id: I05adb7116f1a8b91a9e54dfa70c745f12d7fbf3c