Make Upload/Insert buttons in "Insert media" dialog less confusing
The "Upload" button is now displayed on the left side of the dialog,
which make it clear it's a "secondary" action. "Insert" is on the
right, like all jQuery UI buttons, and the primary action.

Unfortunately jQuery UI doesn't give us a lot of styling options for
buttons or dialogs, even this is a bit of a hack.

Bug: T135917
Change-Id: Ia64229b880653668c11787744c177f6c91900382