Properly disable insert-link button
Use jQuery UI's proper button disabling system, rather than trying
to set attributes manually.

Also remove the alert which used to show when trying to insert a
blank link target, as it now has no use. The functionality was
removed in I3541f438fad9b7e0bf23ead2f2e164e84565d254
when the insert-link button was changed to always disable when an invalid
link was being attempted, but the alert was not noticed at that point
(and nor was this disabling bug).

Bug: T298596
Change-Id: I17c454d6ba336abd0b4eeee0d7166fdf81d25734