Fix unrelated elements flickering during realtime preview
Steps to reproduce:
* Start making changes to a page in the classic editor.
* Use the normal "Show preview" or "Show diff" button at the bottom
  of the editor. The page reloads and now shows a preview/diff at
  the top, above the editor.
* Now start using the realtime preview feature on the right side of
  the editor.

You will notice that the preview/diff above the editor starts to
flicker, but is never updated. (There is no need to update it. The
realtime preview that is updated is the one in the right pane.)

The original code was added via T293347. I believe the flicker
exists since then. It's effectively a misconfiguration. I hope the
comments make this clear.

This is also a direct follow up to T307046 where the diff view was
fixed (it just disappeared before), but flickers since then.

Bug: T293347
Bug: T307046
Change-Id: I481610389cec902268a8fd09f6b0452131752d54