Remove non-helpful whitespace from "Headings" help panel
The help panel is rather cramped, only 125px high. Previewing the
actual margins above and below each heading is not really helpful.
It's also that no other element in the help panel does have this
whitespace.

Also remove a bit of redundancy from some selectors. Saying that
something is for the table "cells" is more than enough. Adding the
"td" doesn't make it better, just longer.

Change-Id: I11bbd5b8ceee709c3a51b20fb06d2315bb4d9316