Sensible limits for resizeable booklet panels
This is relevant for the two colapsible panels with the list of
characters, as well as the help panel. Both have an index on the
right that can be resized. As of now it can be collapsed so much
that it disappears entirely, and enlarged so much that it is larger
than 100%. Both situations look broken and confusing. I think these
are sensible limits.

Also remove the horizontal scrollbar. It appears only when the index
is very narrow, but is especially in the way and not useful in this
situation. Just enlarge the index when you need to see the full
words.

Change-Id: I92d11ddbe4976cf218b30689604ed91b0f288eae