Realtime Preview: Do not display an error when saving a page
Do not show an error in the preview pane if the edit form is
being submitted. The most common error in this case is because
the preview request is aborted, which is not an error. Other
errors could conceivably be in the process of being reported,
but the page is already being reloaded anyway so a flash of
an error message is not very helpful.

Bug: T393077
Change-Id: Ia778e75de3972c52a5b1c562021e89f9d1b5225c