RealtimePreview: Add context and toolbar button creation to constructor
Various parts of the RealtimePreview class use the WikiEditor context
variable, but it was only being set when getToolbarButton() was called.
This meant that other methods couldn't be called, and in init.js the
isScreenWideEnough() method was being called (resulting in an error).

This change moves the context to be injected as a constructor parameter,
and also moves the toolbar-button to be built within the constructor
too. This means that when getToolbarButton is called the button and
context are already available. It does mean we have to fetch the context
from the $textarea's data, but I don't think that's too terrible.

Bug: T394078
Change-Id: I6f3e18610b8dcf39991edcc6c132993526fe23fa