Guard against WikiEditor context not being available
This exits RealtimePreview init early if the WikiEditor context
is not available for any reason. It also switches to the more
common form of 'wikiEditor-context' rather than 'wikiEditorContext'.
Although these are exactly equivalent (data attributes can be
referenced by camel or kebab case), the mixed form is more common
and so easier to grep.

Bug: T395815
Change-Id: I95507edc026360f6b7632a9c7d8c4aa8a311ad8d