dialogs: reset `dialogaction` when opening dialogs
If the insertlink/inserttable dialog was closed by clicking on the cancel button, the keypress no longer works as expected after reopening the dialog because of the previous cancel `dialogaction`. It makes more sense to reset `dialogaction` when opening the dialog.

I am not sure about the insertfile dialog, because the default action is to upload while the user might want to memorize an insert `dialogaction` in many cases.

Bug: T298549
Change-Id: Ifd96a3f3688c4513a13312378b45a4d5fd5d68fe